+++ CountlyConnectionManager.m
@@ -622,7 +622,7 @@ static dispatch_once_t onceToken;
         return;
     }
     
-#if TARGET_OS_IOS || TARGET_OS_TV
+#if TARGET_OS_IOS || TARGET_OS_TV || TARGET_OS_VISION
     if (CountlyServerConfig.sharedInstance.automaticSessionTrackingEnabled && [UIApplication sharedApplication].applicationState == UIApplicationStateBackground) {
         CLY_LOG_D(@"%s App is in the background, 'beginSession' will be ignored", __FUNCTION__);
         return;