+++ CountlyViewTrackingInternal.h
@@ -14,7 +14,7 @@ extern NSString* const kCountlyPreviousEventName;
 extern NSString* const kCountlyVTKeyVisit;
 extern NSString* const kCountlyVTKeyName;
 
-@interface CountlyViewTrackingInternal : NSObject
+@interface CountlyViewTrackingInternal : NSObject <Resettable>
 @property (nonatomic) NSString* currentViewID;
 @property (nonatomic) NSString* previousViewID;